Refactor code for Label and EntityType
