+++ ar/data.py
@@ -1,4 +1,3 @@
-import copy
 import os
 import shutil
 import itertools
@@ -12,7 +11,6 @@ import pandas as pd
 from pandas import DataFrame
 from sklearn.metrics import cohen_kappa_score
 
-from db.model import db, EntityType, Label
 from shared.utils import save_jsonl, load_json, save_json, mkf, mkd
 from db.task import Task, DIR_ANNO, DIR_AREQ
 from db import _task_dir
@@ -20,60 +18,6 @@ from db import _task_dir
 
 ###############################################################################
 # I chose to write it all on disk for now - we can change it to a db later.
-
-def get_or_create(db_session, model, exclude_keys_in_retrieve=None, **kwargs):
-    """Retrieve an instance from the database based on key and value
-    specified in kwargs but excluding those in the exclude_keys_in_retrieve.
-
-    :param db_session: database session
-    :param model: The db model class name
-    :param exclude_keys_in_retrieve: keys to exclude in retrieve
-    :param kwargs: key-value pairs to retrieve or create an instance
-    :return: a model instance
-    """
-    if exclude_keys_in_retrieve is None:
-        exclude_keys_in_retrieve = []
-    read_kwargs = copy.copy(kwargs)
-    for key in exclude_keys_in_retrieve:
-        read_kwargs.pop(key, None)
-
-    instance = db_session.query(model).filter_by(**read_kwargs).one_or_none()
-    if instance:
-        return instance
-    else:
-        instance = model(**kwargs)
-        db_session.add(instance)
-        db_session.commit()
-        logging.info("Created a new instance of {}".format(instance))
-        return instance
-
-
-def fetch_labels_by_entity_type(entity_type_name):
-    """Fetch all labels for an entity type.
-
-    :param entity_type_name: the entity type name
-    :return: all the labels under the entity type
-    """
-    labels = Label.query.join(EntityType).filter(EntityType.name ==
-                                                 entity_type_name).all()
-    return [label.name for label in labels]
-
-
-def save_labels_by_entity_type(entity_type_name, labels):
-    """Update labels under the entity type.
-
-    If entity type doesn't exist, create it first.
-
-    :param entity_type_name: the entity type name
-    :param labels: labels to be saved
-    """
-    logging.info("Finding the EntityType for {}".format(entity_type_name))
-    entity_type = get_or_create(db.session, EntityType, name=entity_type_name)
-    for label_name in labels:
-        label = Label(name=label_name, entity_type_id=entity_type.id)
-        db.session.add(label)
-    db.session.commit()
-
 
 def save_new_ar_for_user(task_id, user_id, annotation_requests,
                          clean_existing=True):
