+++ frontend/labels.py
@@ -1,7 +1,7 @@
 import logging
 
-from ar.data import (
-    fetch_labels_by_entity_type, save_labels_by_entity_type)
+from db.model import (
+    db, fetch_labels_by_entity_type, save_labels_by_entity_type)
 import json
 from flask import (
     Blueprint, request, jsonify, Response)
@@ -15,7 +15,7 @@ bp = Blueprint('labels', __name__, url_prefix='/labels')
 # @login_required
 def fetch_all_labels():
     entity = request.args["entity_type"]
-    labels = fetch_labels_by_entity_type(entity)
+    labels = fetch_labels_by_entity_type(db.session, entity)
     return jsonify(labels)
 
 
@@ -25,14 +25,15 @@ def save_labels():
     data = json.loads(request.data)
     entity_type_name = data['entity_type']
     new_labels = set(data['labels'])
-    labels = set(fetch_labels_by_entity_type(entity_type_name))
+    labels = set(fetch_labels_by_entity_type(db.session, entity_type_name))
     logging.error("Existing labels: {}".format(labels))
     new_labels = new_labels.difference(labels)
     try:
         logging.error("Updating new labels {} for EntityType {}".format(
             str(new_labels), entity_type_name
         ))
-        save_labels_by_entity_type(entity_type_name, list(new_labels))
+        save_labels_by_entity_type(
+            db.session, entity_type_name, list(new_labels))
         msg = "Labels for entity {} have been updated".format(entity_type_name)
         return Response(msg, status=200, mimetype='application/json')
     except Exception as e:
