+++ tests/unit/test_frontend_model.py
@@ -1,6 +1,8 @@
 import frontend
-from ar.data import fetch_labels_by_entity_type, save_labels_by_entity_type
-from db.model import db, Label, EntityType
+from db.model import (
+    db, Label, EntityType,
+    fetch_labels_by_entity_type, save_labels_by_entity_type
+)
 from db.config import TestingConfig
 from flask_testing import TestCase
 
@@ -31,21 +33,21 @@ class FrontEndModelTest(TestCase):
         db.session.commit()
 
     def test_fetch_labels_by_entity_type(self):
-        labels = fetch_labels_by_entity_type(entity_type_name="company")
+        labels = fetch_labels_by_entity_type(db.session, "company")
         for label in test_labels:
             assert label in labels
 
     def test_fetch_labels_for_unknown_entity_type(self):
-        labels = fetch_labels_by_entity_type(entity_type_name="unknown")
+        labels = fetch_labels_by_entity_type(db.session, "unknown")
         assert len(labels) == 0
 
     def test_save_labels_for_existing_entity_type(self):
         new_labels = ["logstic"]
         entity_type_name = "company"
-        save_labels_by_entity_type(entity_type_name, new_labels)
+        save_labels_by_entity_type(db.session, entity_type_name, new_labels)
 
         exisiting_labels = fetch_labels_by_entity_type(
-            entity_type_name="company")
+            db.session, "company")
 
         expected_label_set = set(test_labels)
         expected_label_set.update(new_labels)
@@ -55,10 +57,10 @@ class FrontEndModelTest(TestCase):
     def test_save_labels_for_new_entity_type(self):
         new_labels = ["startup", "x-googler"]
         entity_type_name = "people"
-        save_labels_by_entity_type(entity_type_name, new_labels)
+        save_labels_by_entity_type(db.session, entity_type_name, new_labels)
 
         exisiting_labels = fetch_labels_by_entity_type(
-            entity_type_name=entity_type_name)
+            db.session, entity_type_name)
 
         expected_label_set = set(new_labels)
